Refactor crate into library with bin folder Update CI to build lib, bin, and doc with --no-deps
This puts the binary crate requirements behind a feature flag.
Library users will not need to compile image, png, or the
structopt/clap dependencies. The number of crates needed to
build the lib drops from 67 to 26 crates.
