+++ README.md
@@ -2,7 +2,8 @@
 [![Build Status](https://travis-ci.com/okaneco/kmeans-colors.svg?branch=master)](https://travis-ci.com/okaneco/kmeans-colors)
 [![Crates.io](https://img.shields.io/crates/v/kmeans-colors.svg)](https://crates.io/crates/kmeans-colors)
 
-Calculate the `k` average colors in an image using k-means clustering.
+Calculate the `k` average colors in an image using k-means clustering. The
+crate can also be used as a library.
 
 k-means clustering works by starting with an initial random guess of the `k`
 number of colors in the image called `centroids`. For each step, every pixel in
