+++ src/bin/kmeans_colors/args.rs
