+++ src/bin/kmeans_colors/err.rs
