+++ src/bin/kmeans_colors/filename.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 use std::time::{SystemTime, UNIX_EPOCH};
 
-use super::CliError;
+use crate::err::CliError;
 
 /// Creates a `PathBuf` to save the output filename. Handles the case where user
 /// has specified an output and when there are multiple files that need names.
